scyllaclient: Sanity testing to ensure that host pool rotation works.
When the client can connect via SSH but the API for some reason does
not respond we have to ensure that the next available host is picked.

Iptables is used to simulate the API being down. Note that if the
test program pamics it is possible for the cleanup to fail which
renders a restart of the docker containers necessary.