Scheduler refactoring
This is a complete rewrite of scheduler and runners.

Scheduler is now only place where run status is preserved and managed.
The runner package is removed. The Runner interface is moved to
sched package and now it consists of only Run function.

```go
type Runner interface {
        Run(ctx context.Context, clusterID, taskID, runID uuid.UUID, properties json.RawMessage) error
}
```

The Run function is synchronous and can be stopped by context cancellation.

Fixes #953

Signed-off-by: Michał Matczuk <michal@scylladb.com>