backup: alternative backup location path
This puts node ID and and keyspace to the top to make implementing data
retention easier by grouping metadata and data per node per keysapce.

* This makes the retention metadata removal phase more complex, we need
to iterate keyspace by keyspace. We already do that for all the other
backup steps so it feels a natural thing to do.
* This significantly reduces amount of manifests we need to load
and makes finding them trivial for the data sstables removal.

Signed-off-by: Michał Matczuk <michal@scylladb.com>