cmd/agent/config_test: not relying on files availability on host machine
Default values of agent config for TLS cert and key path pointed
to location which may not contain these files. These location is
not populated during bootstrap neither.
This causes that config validation fails and tests are failing.
This patch changes input configs to use relative paths pointing to files in
repository.