docker: moved to top level docker directory
The docker directory has three different docker setups and serves
as an example setup. It can be started using docker compose
which will bring up one scylla patched with an agent and a Scylla
Manager with it's corresponding database.

The agent will generate new SSL certificates as needed upon startup
to avoid having all the docker containers in the world run the same.