rclone/rcserver: always return Content-Length header
stdlib http server returns this header only for bodies which are shorter
than a few KB. If body is longer, server must take care of adding this
header.
This causes that other side isn't able to return connection back to pool
because it's not sure whether whole body was consumed. That caused
connections to being dropped out from the pool, and new has to be
established.

Fixes #1524