makefile: Extract build targets for binaries
In order to have build step in CI we need to have
targets which just builds binaries. Current dev-agent assumes that
developer environment is running and tries to deploy it to node
containers. Env is not running on CI.

Lint check verifying captital letter at the beginning of the comment now
uses `grep` instead `pcregrep`. install_deps no longer install this
dependency.