testing: add unprivileged user to minio (#1534)
This adds unprivileged user "miniouser" to the minio test environment.
That way we get test environment more similar to the s3 and user can't
create new buckets.

Note that miniouser has only permissions to access buckets with
"backuptest-" and "manual-" prefixes.

Also default region for the minio server is set to us-east-1. But
default minio behavior is to allow operations on default region when
region is not specified. Explicitly specified region that is different
from the one of the bucket will produce errors.

For testing purposes we are creating buckets by creating directories
and this puts buckets into the default region of the server.

Fixes #1534