docker: makefile for building manager and agent
To build and tag a given version you can run:

  make all

This will create images for both the Manager server and the Agent
and deploy them to the local registry.
The environment variable SCYLLA_MANAGER_VERSION decides which
version of the Manager and Agent is used and defaults to the value
set in the .env file.