cmd/agent: replace rclone gc with FreeOSMemory (#1675)
We are replacing rclone GC call with debug.FreeOSMemory call
to release the memory back to the OS.

Cleanup will be called after upload and at the end of the backup.
Cleanup no longer calls gc.

FreeOSMemory call is enabled on the agent for debugging purposes and
it will not be called during normal execution because we established
that runtime is doing a good job at that.

Fixes #1675