rclone/rcserver: override operations/list to stream results (#1657)
Rclone version of operations/list is loading all of the listed files
into memory before sending them as a response to the wire.

This change is overriding this operation to stream output back to the
caller as soon as results are available.

If there are less than 1000 items to list streaming will not be used.

If there is an error during streaming error field will be added to the
listing response.

Fixes #1657

Signed-off-by: Michał Matczuk <michal@scylladb.com>