service/backup: restart upload on agent restart (#1706)
With introduction of long poling, job wait logic was broken and
restarting an agent during backup process would fail the backup task.

With this fix job info response was changed to return nil job stats when
job can't be found instead of error.
This results in proper signal that job restart is needed.

Fixes #1706