service/backup: fix stage data race (#1894)
Stage field of the Run object is not safe for concurrent use.
But it's accessed concurrently. It is updated on each backup step,
and read on each metrics update tick.

To isolate run use between metrics and backup task execution
run information for metrics will be loaded from the database on
each tick.

Fixes #1894