service/backup: Allow backups when some nodes are downed or drained (#1551, #1782)
This patch makes it possible to run backup when there are enough replicas alive.
A node down or nodes down will not fail the backup if there are replicas available for each token.

GetTarget is changed to do the live node selection, Backup uses the host info passed in Target.

Fixes #1551
Fixes #1782