scyllaclient: use native_transport_port_ssl for TLS CQL connections
Due to undesired behavior of Scylla API we cannot tell if
native_transport_port_ssl was configured by user or it's a not being used
default value. To workaround this we try to establish connection
to native_transport_port_ssl with 1s timeout to check whether it's open
or not. If it's not open, we fall back to native_transport_port.

Workaround is not applied when client encryption is disabled.

Healthcheck iterations, and first CQL connection to managed cluster are
affected.

Reference: https://github.com/scylladb/scylla/issues/7206

Fixes #2137