service/repair: Gracefully handle table and keyspace deletion during repair (#1161)
As repair may take a lot of time we need to ignore tables that were deleted during a repair.
If delete happens in init phase is will error the repair and trigger retry from scheduler.
If delete happens during the repair process it will be handled gracefully.
All token ranges for the deleted table would be assumed successfully repaired without ever sending them to Scylla.

ScyllaClient is extended with TableExists function.

Worker is changed to check for table removal if
- Repair scheduling fails with 400 HTTP status code
- Repair status fails with 400 HTTP status code
- Scylla repair fails

Generator is changed to mark and ignore the deleted tables.

Fixes #1161