scheduler: add generic purpose scheduler
The scheduler package is designed after Quartz project.
It borrows the idea of managing keys and a separation of task properties from the run queue.
Please find the mapping of Quartz core components to scheduler below

- [Job](https://www.quartz-scheduler.org/api/2.4.0-SNAPSHOT/org/quartz/Job.html) - RunFunc
- [JobKey](https://www.quartz-scheduler.org/api/2.4.0-SNAPSHOT/org/quartz/JobKey.html) - Key
- [JobDetail](https://www.quartz-scheduler.org/api/2.4.0-SNAPSHOT/org/quartz/JobDetail.html) - Properties
- [Trigger](https://www.quartz-scheduler.org/api/2.4.0-SNAPSHOT/org/quartz/Trigger.html) - Trigger

The interfaces are decluttered and translated to idiomatic Go.
Renames and customizations are done to better fit the current scheduler infrastructure (pkg/service/scheduler).

Please check Scheduler docs for details.