repair: add support for long polling
Long polling repair status checks are added for Scylla nodes that
support it.

Feature was tested manualy on Scylla 4.3.rc0.

Integration tests needs to be updated to the Scylla 4.3.0 once it's out.

service/backup: refactored long polling

Changed how long poling timeout configuration is structured in backup.
Long polling timeout is now provided to the client as a parameter.