service/healthcheck: rest, report status error if rtt=0
This adds the missing case in a switch and makes error handling of REST, CQL and alternator fully symmetric.
It does not really "fix" the issue #2973 because ERROR is reserved for a precondition failure.
It's valid to have REST DOWN and CQL ERROR due to a failure to get CQL addresses from the Agent REST API...

Fixes #2973