service/scheduler: resolver, dumbify resolver
The resolver was built with a trie to be able to match prefixes of task IDs.
This is replaced with two maps, cluster ID + task name to task ID map, and task ID to task info map.
The resolver API remains more or less the same as the prefix functionality was not used.

The rationale behind it is that it's better to have your own short name instead of a prefix or a random ID.
When working with ID you would copy it anyway.