scyllaclient: retry increase t/o bugfix
Without this patch the timeout increase is linear adding 2 * config.timeout after every fail due to a timeout.
It was not intended, and tests also worked thanks to a lucky timings.

This patch fixes the exponential timeout behaviour.
It adds MaxTimeout (default 1h) config option to scyllaclient.Config, it is propagated to retryConfig to limit the effective max timeout.

Fixes #3011