scyllaclient: t/o per batch on RcloneListDirIter
Instead of passing a context into the HTTP request with a 12 hour
timeout, we now set a 5 minute timeout but only enforce it per batch of
files. This is done by waiting for the next JSON decode from the HTTP
request body and selecting between that and a 5 minute timeout each time
we process a new batch.

Adds noTimeout function for scyllaclient contexts. This should only be
used for RcloneListDirIter.