scheduler/window: Account for DST changes
Previously we were calculating execution times by simply adding a
time.Duration to 00:00 on the execution day. Most of the times this
works, however on days where DST begins or ends it does not. For
example 21:00 is not always 21 hours after 00:00, sometimes it can
be 20 or 22 hours after.

In order to account for this we are now creating a new time using the
hours from duration and the location instead of simply adding it.

Also fixed an issue with a test that does not account for DST changes,
sometimes EST and CET are not 6 hours apart but instead 5 or 7 because
the clocks change on different dates.

fixes #3082