service/scheduler: schedule one off tasks on resume
If a task is created with a window but not any type of schedule it
should begin execution when the window begins. In the case where the
scheduler is suspended or stopped before the execution that was never
happening. This makes sure we check if a task created in that way gets
scheduled on a resume if the task has never run.

fixes #3083