command/suspend: validate suspend command
When creating or updating a suspend command, make sure that start-tasks
is not supplied without a duration. This has to be checked both when
creating or updating an actual suspend task, as well as when suspending
immediately as a one-off when there is no task to validate.

The duration in `props` can be either a `string` or a `duration.Duration`
depending on whether the duration is being updated in the command or
unchanged from the existing task, so there are a handful of validation
cases needed to check for.

fixes #3080