managerclient: don't limit status column
As far as I can tell this limit really didn't bring any benefit to begin
with. Every status is almost as short as the table header "Status"
anyway and the other columns in the table are much longer and more
likely to cause the lines to wrap. There is also only a 1-2 width window
where this could actually be beneficial, and in other cases it leads to
every status showing as "…" which can cause confusion.

fixes #3065