service/backup: iteratively load manifest indexes
When purging manifests there is no reason to load the other manifest
data, as only the indexes are used, so this skips directly to the index
data and iteratively decodes the Indexes from the manifest json.

Other uses of the manifests still load everything at once, as they are
actually using the full manifest object.

relates to #3116