pkg/scheduler: Rename subtests of TestWindowParse
This commit is required to make the module go-gettable.

Name of the tests are used to generate golden files and some characters
are not allowed to be presented in the filenames of a Go module.

Previous version has `:` character in the filenames that prevents
downloading the whole module with `go get` command.

See:
- https://go.dev/ref/mod#zip-path-size-constraints