pkg/testutils: Add check for golden file name creator
This commit is required to make future additions of
the golden files compatible with the Go module restrictions
on file names.

Every existing file returned by a pattern `*.golden.*` has
been checked against the `module.CheckFilePath()` to ensure
that all existing ones are compliant with the Go module
restrictions.

This commit adds a dependency on the golang.org/x/mod package.

See:
- https://go.dev/ref/mod#zip-path-size-constraints