command/flag: added separate type for StartDate which parses 'now' into zero value
In case of '--start-date now' we want to calculate 'now' at the time of task scheduling, NOT when parsing command arguments. Without that it's difficult to distinguish tasks scheduled with --start-date that is set in the past and tasks scheduled with 'now'.
This commit unifies the way in which SM handles tasks with '--start-date now' and tasks without this flag set.