fix(repair): schedule and run repair on single node cluster
SM 3.2 changed repair param validation so that scheduling and running repair of a single node cluster is treated as an error. Although this is acceptable in terms of real users, the ability to schedule/run repair on a simple single node cluster is useful for tests and some experiments.

Fixes #3649