refactor(swagger): drop unused API
"/storage_service/describe_ring/" returns the token range of a
random keyspace. this API is never used in production. what is used
is is cousin which accepts a mandatory keyspace path parameter, like
"/storage_service/describe_ring/{keyspace}", so let's drop it in
scylla-manager, we will drop this API in scylla as well.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>