fix(client): check the initial host first
The fix for https://github.com/scylladb/scylla-manager/issues/3707 includes extending the cluster kept in the DB with the 'host'
property, which represents the initial host used when adding a cluster to Scylla Manager. Scylla-operator passes the DNS name here,
making it immune to the ephemeral IPs in the Kubernetes environment. This commit adds the initial host as the first host to query
in order to discover node IPs.