fix(restore_test): assume 'repair' to be default tombstone_gc mode
Up to this point we didn't specifically set tombstone_gc mode in tests and expected it to be 'timeout'. Since tablet tables need to have tombstone_gc 'repair', we should change the default expectation to it.
Note that this 'assumption' should be removed in the future in favor of something more robust.