feat(restore): new indexing and batching approach
The idea is to first index all files to be restored, so that we can create better batches. Restore workload is aggregated first by table, then by remote sstable dir.

From batching we expect:
- batch contains X*shard_cnt sstables
- batch contains similarly sized sstables
- batch is created from any manifest/table - no waiting for manifest/table restore to finish
- workload across different nodes is evenly distributed
- --batch-size is ignored, batches are aiming to have size equal to 5% of expected node workload

New flag --table-parallel - it allows for running multiple download and l&s jobs from the same node (no documentation as it might not be exposed later on).