feat(backup): allow for not backing up schema
Previous commit made it impossible to perform backup of cluster that
restores schema from desc schema without CQL creds.
In order to make up for that, backup properties are extended with
DontBackupSchema which can be deliberately set in order to skip
schema backup and allow to back up user data when no CQL creds were set.