feat(restore_test): test restore pause/resume
This commit adds runPausedRestore function which allows to run
restore which is going to be interrupted every specified interval.
Pausing after some arbitrary amount of time might seem flaky,
but we can't always rely on scyllaclient hooks for pausing restore,
as we are biased to where we put them.

This commit also alters runPausedRestore to include views in schema
and the backup, so that it can test for #4037.