feat(restore): batch, scale batch size with shard count
This was the initial design, that was changed during implementation by mistake.
We should always send shard_cnt * --batch-size sstables in a single restore job,
even when user wants to restore into a running cluster. For this case, it should
be enough to set --parallel=1 and --batch-size=1 for a slow running restore.