feat(rcserver): implement rate limit/transfers change on upload/download
This commit allows for greater control over transfers
and rate limit as a part of the upload/download
rclone API used by backup/restore tasks.
It also makes the changes more persistent in case
of agent restarts.

It also adds the core/transfers endpoint that
can be used outside the upload/download rclone API
in order to set the amount of transfers for the following rclone API calls.