fix(backup_test): remove partial upload check from TestBackupResumeIntegration
This check isn't implemented correctly, as starting from 27b1d233,
file names are not stored in the SM DB. It means that 'sfs' was
always empty.
Nevertheless, the partial upload check could still fail when
's3f' was empty as well. This test does not implement any breakpoint
to ensure that backup is paused after transferring at least a single file
(it just waits for the transfers to start).
That's why this check was flaky and didn't do anything meaningful,
so it is removed from the test.