feat(restore_test): extend TestRestoreTablesPreparationIntegration with transfers
This way this test also checks transfers before and after backup.
It also checks transfers before, in the middle, when paused,
when resumed, and after restore.
This required some changes to the test like swapping
src and dst cluster (increase batch count) and hanging on LAS instead of
copy paths (transfer change is applied as a part of copy paths).