feat(swagger): agent, add bandwidth_rate and transfers to upload/download APIs
This is a groundwork needed for #3947 and #3948.
Initially, the idea was to use dedicated endpoints to set
the bandwidth_rate and transfers count used by all following
rclone API calls (4c140768).
This does approach does not work well with agent restarts,
as the changes made by dedicated endpoints are not preserved
after the restart.
In order to change that, we add bandwidth_rate and transfers
to rclone API calls responsible for uploading/downloading files.
This way, each call to rclone API will validate that rclone server
uses required amount of transfers (or set the amount if needed).