feat(restore): support max transfers (0) and make it default
Our experiments showed that the fastest file download
was achieved for transfers=2*2*node_shard_cnt. In order to make
it easier for the user to use, a new, special value
of transfers=0 will take care of that.
It is set as the default, because we aim to make
not configured restore as fast as possible.