feat(restore_test): extend TestRestoreTablesPreparationIntegration with rate limit
This way this test also checks rate limit before and after backup.
It also checks transfers before, in the middle, when paused,
when resumed, and after restore.

This commit also extends the test to change transfers
and rate limit values when restore is paused, so that
it validates that they are correctly re-set during
restore data stage.