Restore improvement: transfers (#4054)
* chore(go.mod): bump SM sub packages versions

* feat(rcserver): implement rate limit/transfers change on upload/download

This commit allows for greater control over transfers
and rate limit as a part of the upload/download
rclone API used by backup/restore tasks.
It also makes the changes more persistent in case
of agent restarts.

It also adds the core/transfers endpoint that
can be used outside the upload/download rclone API
in order to set the amount of transfers for the following rclone API calls.

* feat(scyllaclient): add RcloneSet/GetTransfers to client

This commit exposes the /rclone/core/transfers agent endpoint via agent client.

* feat(scyllaclient): set transfers in RcloneCopyPaths

This commit requires changes in the usage of RcloneCopyPaths in restore pkg.
It also adds Transfers field to restore Target with the default value -1
meaning that transfers will be set to the amount specified in agent
rclone server config.

* feat(restore): support max transfers (0) and make it default

Our experiments showed that the fastest file download
was achieved for transfers=2*2*node_shard_cnt. In order to make
it easier for the user to use, a new, special value
of transfers=0 will take care of that.
It is set as the default, because we aim to make
not configured restore as fast as possible.

* feat(command/restore): add --transfers flag

This commit allows user to control the amount of rclone transfers used during restore.
Moreover, during development days of restore task, the default value of --parallel flag
was changed from 1 to 0, but we forgot to change it in the flag documentation.

* feat(backup): add Transfers to Target

This is the first step to control transfers in the context of backup.

* feat(scyllaclient): set transfers in rcloneMoveOrCopyDir

This commit requires changes in the usage of RcloneMoveDir in backup pkg.

* feat(command/backup): add --transfers flag

This commit allows user to control the amount of rclone transfers used during backup.

* feat(restore_test): extend TestRestoreTablesPreparationIntegration with transfers

This way this test also checks transfers before and after backup.
It also checks transfers before, in the middle, when paused,
when resumed, and after restore.
This required some changes to the test like swapping
src and dst cluster (increase batch count) and hanging on LAS instead of
copy paths (transfer change is applied as a part of copy paths).