feat(restore): allow for unpinning agent from CPU (#4073)
* chore(go.mod): bump sub pkg versions

* feat(agent): allow for setting multiple CPUs in config

This commit makes it possible to specify multiple
CPUs in 'scylla-manager-agent.yaml' config to which
agent will be pinned. The 'cpu' field in config
now allows for both single int and array of ints
values. The default behavior remains the same.

* refactor(agent): extract cpu pinning logic to separate function

Except for the extraction, this commit also increases log
level (from DEBUG to INFO) of message about missing cpuset file.
It also makes this function return an error, which is going
to be useful when this function is going to be used inside
the cpu pinning agent endpoints.

* feat(agent): implement cpu pinning endpoints

This commit implement 3 cpu pinning endpoints
allowing to query (GET), pin (POST) and
unpin (DELETE) agent cpus.

* feat(scyllaclient): implement cpu pinning methods

* feat(restore): optionally unpin agent from CPUs

This commit adds UnpinAgentCPU field to Target.
It optionally allows to unpin agent from CPUs
for the time of the restore.

Fixes #3951

* feat(restore): always enable auto compaction after restore

This should be done just because of safety.
It prevents rare edge cases like:
- restore runs with --allow-compaction=false
- restore is paused
- restore is modified with --allow-compaction=true
- restore is resumed

* feat(command/restore): add --unpin-agent-cpu flag

This commit allows user to control whether
agent should be pinned to CPUs during restore.

* feat(restore_test): extend TestRestoreTablesPreparationIntegration with cpu pinning

This way this test also checks cpu pinning before and after backup.
It also checks cpu pinning before, in the middle, when paused,
when resumed, and after restore.