feat(restore): add bandwidth metrics (#4081)
* refactor(restore): separate methods for updating metrics/progress

This should make it easier to see what is updated where and when.

* feat(metrics): restore, add bandwidth metrics

They are really useful for evaluating restore performance.

* feat(restore): set download/stream bytes/duration metrics

It's useful for checking/tracking restore performance.

Ref #4042

* fix(restore): don't initialize metrics twice

This was a left-over from the PR introducing
indexing (14aef7be). It also initialized metrics
as a part of the indexing procedure, but it
forgot to remove the previous metrics initialization
from the code.

* fix(restore): use backup bluster ID in remaining_bytes metric

There was a confusion about which cluster ID should
be used for labeling remaining_bytes metric.
When setting remaining_bytes, we used backup cluster ID,
but when decreasing, we used restore cluster ID.
Backup cluster ID should be used in both places
as this metrics describes how many bytes from
which place are yet to be restored. Since we use backup
cluster DC, node ID, etc., we should also use backup
cluster ID.