fix(healthcheck): correctly label nodes removed from metrics
Commits a75df89c and 3cdd20dc added and started setting
new labels (dc/rack) to the healthcheck metrics.
Unfortunately, those labels weren't taken into consideration
when removing labels of removed clusters/nodes.
This resulted in not removing those metrics until SM restart.

This commit fixes this issue and brings a small refactor
to the way in which healthcheck labels are applied,
so that it's more difficult to make such mistake in the future.

Fixes #4017