feat(schema): new restore run progress
We need to extend run progress with Scylla task ID.
Moreover, we need it to be a part of the clustering key
Otherwise, we won't be able to store multiple run progresses
with the same remote sstable dir and host in the DB.
That's why we need to re-create the table.
It also allows us to rename manifest path and versioned
progress to more suitable names.
The downside is that users will lose their restore progress
history on upgrade, but this shouldn't harm anybody
if communicated properly.