Make restore progress available even when cluster is not (#4196)
* feat(schema): extend view with build status

We need to store view build status because we want
to display it during in sctool progress.

Ref #4191

* refactor(restore): move view build status to view

Since build status is a part of the view definition
in SM DB schema, we should also reflect it in the code.
This also allows us to get rid of the ViewProgress struct.

* feat(restore_test): test view progress after cluster is unavailable

This test checks for #4191.

* fix(restore): store view build status in SM DB instead of querying it in progress

This way there is no problem with getting sctool progress
even when the cluster is no longer available.
This makes TestRestoreTablesProgressIntegration pass.

Fixes #4191

(cherry picked from commit f4426163de10c0a6aa4afdee201a3adf981552e3)