feat(swagger): scylla-manager, move towards just restored bytes (#4214)
In overall/keyspace/table progress we were reporting both
downloaded and restored bytes. There are two problems with it:

1. Users does not really care for the distinction between
downloaded and restored bytes - they just want to know the
amount of bytes that have been successfully restored.

2. With new native Scylla API restore, we no longer separate
restoration into download/load&stream stages.

That's why we should move towards reporting just the amount
of successfully restored bytes, and not keep on adding another
fields to our API for each new type of restore.

This commit:
- Adds restored_bytes/duration to host progress, since it was
lacking there. It does not deprecate the downloaded/streamed
fields because they were just added in the previous minor release,
so that would be confusing. Instead, for now we allow for host
progress to contain detailed operation bandwidth information.
- Deprecates downloaded bytes in overall/keyspace/table progress,
as we want to move towards observing just the restored bytes.

A general note is that we should use metrics for monitoring
performance, not the sctool progress cmd, but since we already
started going in this direction, we need to support it.