fix(restore_test): skip unintended CQL schema tests
This commit adds skipCQLSchemaTestAssumingSSTables helper method
which allows to skip the CQL schema tests which were intended
just for the SSTAble schema restore. Example of such tests:
- restoreWithAgentRestart
- restoreWithResume
- restoreWithVersions
- restoreAlternator

Those tests were "passing" before by accident and that's
why they weren't skipped in the past. They are also not
that interesting for CQL schema restore, which just
applies saved CQL statements.

It also rephrases some comments so that it's clear
when a schema test should be skipped and for what
reason.

Fixes #4246