feat(repair): don't disable tablet load balancing from Scylla 2025.1.0
Scylla 2025.1.0 is introducing new tablet repair API which does not
require stopping tablet load balancing during the repair.
Unfortunately, host filtering for this API will be added in 2025.1.1,
so we can't use this API for partial tablet tables repair.

As Scylla 2025.1 is aiming to utilize ~90% of disk, disabling tablet
load balancing for an extended period of time (e.g. repair) on a
running cluster might result in some nodes running out of disk space.

Since skipping repair of some tablets does not lead to data resurrection
(because of tombstone_gc mode repair), it's better to not stop tablet
load balancing during repair with Scylla 2025.1.0 even when partially
repairing tablet table with the old API.

Ref #4273